Before datacenter update
Updated
	Storage nodes are not cleared when removed. Only the status is changed

Next to do:
	Update data center removed. Need to be kept in the contract after removed
	with status updated to `removed`
