+++ VDA-StorageNode-Registry/scripts/contract-address.json
@@ -1,13 +1,14 @@
 {
   "hardhat": {
     "date": "2023/12/05",
-    "diamond": "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6",
-    "tokenAddr": "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853",
+    "diamond": "0x8A791620dd6260079BF849Dc5567aDC3F2FdC318",
+    "tokenAddr": "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6",
     "DiamondCutFacet": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512",
     "DiamondLoupeFacet": "0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0",
     "OwnershipFacet": "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
     "VDAVerificationFacet": "0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9",
     "VDADataCenterFacet": "0x5FC8d32690cc91D4c39d9d3abcBD16989F875707",
-    "VDAStorageNodeFacet": "0x0165878A594ca255338adfa4d48449f69242Eb8F"
+    "VDAStorageNodeFacet": "0x0165878A594ca255338adfa4d48449f69242Eb8F",
+    "VDAStorageNodeManagementFacet": "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853"
   }
 }
